refactor homepage, add store page, caching network image, add some icon
