+++ lib/main.dart
@@ -7,6 +7,7 @@ import 'package:unimarket/utilities/constants.dart';
 import 'controller/home_provider.dart';
 import 'controller/product_provider.dart';
 import 'controller/profile_provider.dart';
+import 'controller/store_provider.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -31,6 +32,7 @@ class MyApp extends StatelessWidget {
         ChangeNotifierProvider(create: (_) => HomeProvider()),
         ChangeNotifierProvider(create: (_) => ProductsProvider()),
         ChangeNotifierProvider(create: (_) => ProfileProvider()),
+        ChangeNotifierProvider(create: (_) => StoreProvider()),
       ],
       child: MaterialApp(
         debugShowCheckedModeBanner: false,
