+++ lib/screens/profile.dart
@@ -1,7 +1,10 @@
 import 'dart:math';
 
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_svg/flutter_svg.dart';
 import 'package:unimarket/screens/auth/login.dart';
+import 'package:unimarket/screens/store/store.dart';
 import 'package:unimarket/utilities/constants.dart';
 
 import '../controller/auth_provider.dart';
@@ -28,9 +31,12 @@ class _ProfileState extends State<Profile> {
                 children: [
                   CircleAvatar(
                     radius: 48,
-                    backgroundImage: NetworkImage(
+                    backgroundImage: CachedNetworkImageProvider(
                       'https://picsum.photos/id/$randomNumber/200/200',
                     ),
+                    // NetworkImage(
+                    //   'https://picsum.photos/id/$randomNumber/200/200',
+                    // ),
                     child: Align(
                       alignment: Alignment.bottomRight,
                       child: Icon(
@@ -64,19 +70,39 @@ class _ProfileState extends State<Profile> {
             formSpacer,
             const Divider(),
             ListTile(
-              leading: const Icon(Icons.sell_outlined, color: Colors.black),
+              leading: SvgPicture.asset(
+                'assets/icons/edit.svg',
+                width: 20,
+                height: 20,
+              ),
               title: const Text('Request Sebagai Penjual'),
               trailing: const Icon(Icons.chevron_right_rounded),
               onTap: () {},
             ),
             ListTile(
-              leading: const Icon(Icons.store, color: Colors.black),
+              leading: SvgPicture.asset(
+                'assets/icons/shop.svg',
+                width: 20,
+                height: 20,
+              ),
               title: const Text('Kelola Toko (\'Tokomu\')'),
               trailing: const Icon(Icons.chevron_right_rounded),
-              onTap: () {},
+              onTap: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => const Store(),
+                  ),
+                );
+              },
             ),
             ListTile(
-              leading: Icon(Icons.logout_rounded, color: Colors.red[900]),
+              leading: SvgPicture.asset(
+                'assets/icons/exit.svg',
+                width: 20,
+                height: 20,
+                color: Colors.red[900],
+              ),
               title: const Text('Logout'),
               trailing: const Icon(Icons.chevron_right_rounded),
               onTap: () async {
