+++ lib/screens/store/my_orders.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class MyOrders extends StatefulWidget {
+  const MyOrders({Key? key}) : super(key: key);
+
+  @override
+  State<MyOrders> createState() => _MyOrdersState();
+}
+
+class _MyOrdersState extends State<MyOrders> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Column(),
+    );
+  }
+}
