+++ pubspec.yaml
@@ -43,6 +43,7 @@ dependencies:
   email_validator: ^2.1.17
   flutter_staggered_grid_view: ^0.6.2
   badges: ^3.1.1
+  cached_network_image: ^3.2.3
 
 dev_dependencies:
   flutter_test:
